+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+- Parse files correctly if the user has spefied globs without extensions.
+
 ## [0.9.0] - 2022-05-11
 ### Fixed
 - Ignore parse errors and print them to STDERR